Some general clean up and file restructuring
