+++ client/package.json
@@ -20,6 +20,7 @@
     "react-scripts": "^3.0.1",
     "react-spinners": "^0.5.3",
     "react-tabs-redux": "^4.0.0",
+    "react-underline-tabbar": "^1.1.7",
     "react-web-tabs": "^1.0.1",
     "reactstrap": "^7.1.0",
     "redux": "^4.0.1",
