+++ client/src/App.js
@@ -1,13 +1,10 @@
 import React, { Component } from 'react';
-
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 
 // Views
 import SignInView from './views/SignInView'
 import HomeView from './views/HomeView'
 import ProfileView from './views/ProfileView'
-import VizTestView from './views/VizTestView'
-
 
 class App extends Component {
   render() {
@@ -17,7 +14,6 @@ class App extends Component {
           <Route exact path='/' component={SignInView} />
           <Route path='/home' component={HomeView} />
           <Route path='/profile' component={ProfileView} />
-          <Route path='/viz' component = {VizTestView} />
           <Redirect to='/' component={SignInView} />
         </Switch>
       </Router>
