+++ client/src/index.js
@@ -9,7 +9,7 @@ import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux'
 import { createStore, applyMiddleware } from 'redux'
 import thunkMiddleware from 'redux-thunk'
-import rootReducer from './reducers'
+import rootReducer from './redux/reducers'
 
 let createStoreWithMiddleware = applyMiddleware(thunkMiddleware)(createStore)
 const store = createStoreWithMiddleware(rootReducer)
