+++ client/src/views/ProfileView.js
@@ -1,24 +1,23 @@
 import React, { Component } from 'react'
-
 import ProfileContentView from './ProfileContentView'
 import ProfileBanner from '../components/ProfileBanner'
 import ProfileNavBar from '../components/ProfileNavBar'
-import UnderlinedTabs from '../components/UnderlinedTabs'
 
 const TabsContainerStyle = {
     fontSize: '15pt',
     marginTop: '8vh',
+    height: '8vh',
     paddingLeft: '12vh',
 }
 
+// Tab title : content view
+const tabs = {
+    "Details": <ProfileContentView />, 
+    "Related": <div></div>
+}
+
 class ProfileView extends Component {
-    constructor(props) {
-        super(props)
-        this.state = {activeTab: 'Details'}
-    }
-
-    render() {
-
+    render() {        
         return (
             < div >
                 <ProfileBanner />
