--- client/src/views/VizTestView.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class VizTestView extends Component {
-    render() {
-        return (
-            <div>
-                This is the viz test page
-            </div>
-        )
-    }
-}
-
-export default VizTestView
